Moved test file to different folder, moved some stuff to main.cpp. Will need to add tasks to scheduler for functionality, nothing pops up  on serial monitor yet. can't use serial in interrupts so look out for that
